refactor: extract shared exlm-config-utils from premium-learning-utils
Moves the private getExlmConfig/fetchExlmConfig helpers into a shared
module and adds isDomainAllowed(), so the comma-separated domain-list
check isn't duplicated. No behavior change for Premium Learning.
