+++ scripts/utils/premium-learning-utils.js
@@ -1,5 +1,6 @@
 import { setCookie, getCookie, deleteCookie } from './cookie-utils.js';
 import isFeatureEnabled from './feature-flag-utils.js';
+import { isDomainAllowed } from './exlm-config-utils.js';
 
 const LEARNER_TOKEN_COOKIE = 'alm_access_token';
 const LEARNER_USER_ID_COOKIE = 'alm_user_id';
@@ -7,36 +8,11 @@ const DEFAULT_EXPIRES = 86400;
 const PL_ELIGIBILITY_TIMEOUT_MS = 10000;
 const isUEMode = window.hlx?.aemRoot || window.location.href.includes('.html');
 
-// Lazily initialised — null until the first getExlmConfig() call.
-// Deferred so window.hlx.codeBasePath is guaranteed to be set by setup() before the URL is read.
-let exlmConfigPromise = null;
-
 // Two separate singletons for the two mutually exclusive auth modes (UE Author vs production).
 // UE/non-UE is an immutable page-level constant, so each promise is set at most once per load.
 // Sign-out calls window.adobeIMS.signOut() which causes a full page reload, resetting both.
 let plAuthPromise;
 let plAuthAnonymousPromise;
-
-function fetchExlmConfig() {
-  if (!exlmConfigPromise) {
-    try {
-      exlmConfigPromise = fetch(`${window.hlx.codeBasePath}/exlm-config.json`, {
-        signal: AbortSignal.timeout(5000),
-      })
-        .then((res) => (res.ok ? res.json() : { data: [] }))
-        .then(({ data = [] }) => new Map(data.map(({ key, value }) => [key, value])))
-        .catch(() => new Map());
-    } catch {
-      exlmConfigPromise = Promise.resolve(new Map());
-    }
-  }
-  return exlmConfigPromise;
-}
-
-async function getExlmConfig(key) {
-  const config = await fetchExlmConfig();
-  return config.get(key) ?? null;
-}
 
 export function getPLAccessToken() {
   return getCookie(LEARNER_TOKEN_COOKIE);
@@ -152,14 +128,7 @@ export async function isPLEligible(signedIn = null, timeoutMs = PL_ELIGIBILITY_T
     return verifyPLAuth(timeoutMs, false);
   }
   if (signedIn === false) return false;
-  const rawDomains = await getExlmConfig('plAllowedDomains');
-  const plAllowedDomains = rawDomains
-    ? rawDomains
-        .split(',')
-        .map((d) => d.trim())
-        .filter(Boolean)
-    : [];
-  if (!plAllowedDomains.includes(window.location.hostname)) return false;
+  if (!(await isDomainAllowed('plAllowedDomains'))) return false;
   return verifyPLAuth(timeoutMs, false);
 }
